always allow 100% size

+++ code/modules/vore/resizing/resize_vr.dm
@@ -89,19 +89,20 @@
 //#define USE_DIRECT_MULTIPLIERS //Uncomment to use old resize calculations.
 /mob/living/proc/resize(var/new_size, var/animate = TRUE, var/uncapped = FALSE, var/ignore_prefs = FALSE, var/aura_animation = TRUE)
 	if(!uncapped)
-		#ifdef USE_DIRECT_MULTIPLIERS
-		new_size = clamp(new_size, RESIZE_MINIMUM, RESIZE_MAXIMUM)
-		#else
-		var/size_diff = ((runechat_y_offset() / size_multiplier) * new_size) // This returns 32 multiplied with the new size
-		var/size_cap = world.icon_size * (RESIZE_MAXIMUM+(ishuman(src)?0:0.5)) //Grace for non-humanoids so they don't get forcibly shrunk.
-		if(size_diff - size_cap  > 0)
-			var/real_diff = size_cap / size_diff // Returns our diff based on the offset to world size
-			new_size *= real_diff // Applies our diff to the new size
-			new_size = clamp(new_size, RESIZE_MINIMUM, RESIZE_MAXIMUM) //If the sprite is below 32, we clamp it to only go to the resize max.
-		#endif
-		var/datum/component/resize_guard/guard = GetComponent(/datum/component/resize_guard)
-		if(guard)
-			qdel(guard)
+		if(new_size != RESIZE_NORMAL) //If our new_size is RESIZE_NORMAL, always allow it without doing extra math. 100% is ALWAYS allowed.
+			#ifdef USE_DIRECT_MULTIPLIERS
+			new_size = clamp(new_size, RESIZE_MINIMUM, RESIZE_MAXIMUM)
+			#else
+			var/size_diff = ((runechat_y_offset() / size_multiplier) * new_size) // This returns 32 multiplied with the new size
+			var/size_cap = world.icon_size * (RESIZE_MAXIMUM+(ishuman(src)?0:0.5)) //Grace for non-humanoids so they don't get forcibly shrunk.
+			if(size_diff - size_cap  > 0)
+				var/real_diff = size_cap / size_diff // Returns our diff based on the offset to world size
+				new_size *= real_diff // Applies our diff to the new size
+				new_size = clamp(new_size, RESIZE_MINIMUM, RESIZE_MAXIMUM) //If the sprite is below 32, we clamp it to only go to the resize max.
+			#endif
+			var/datum/component/resize_guard/guard = GetComponent(/datum/component/resize_guard)
+			if(guard)
+				qdel(guard)
 	else if(has_large_resize_bounds())
 		if(is_extreme_size(new_size))
 			AddComponent(/datum/component/resize_guard)
